chore: Move Avatar component to UI directory
