--- packages/webapp-libs/webapp-core/src/components/avatar/avatarFallback/index.ts
@@ -1,1 +1,0 @@
-export * from './avatarFallback.component';
