--- packages/webapp-libs/webapp-core/src/components/avatar/avatarImage/index.ts
@@ -1,1 +1,0 @@
-export * from './avatarImage.component';
