--- packages/webapp-libs/webapp-core/src/components/avatar/index.ts
@@ -1,3 +1,0 @@
-export * from './avatar.component';
-export * from './avatarFallback';
-export * from './avatarImage';
